refactor: update count down section
